moved spawner package into control script
